refactor: revert back to non package managed
Breaking everything into packages turned out to be a bad move. It was difficult to manage and resulted in constant build crashes and more time was spent trying to get things to work then actually building. Reverting back to a simplier time. We will reconsider how to approach conversion to lambda functions.